test and fix bugs

+++ config/config.go
@@ -137,8 +137,8 @@ type LogOptions struct {
 
 type TimingMatcher struct {
 	RoundOrdersCount int
-	Duration         int
-	DelayedNumber    int
+	Duration         int64
+	DelayedNumber    int64
 }
 type MinerOptions struct {
 	RingMaxLength           int    `required:"true"` //recommended value:4
